Addressing comments
