+++ src/software/jetson_nano/display/screens/chip_and_kick_screen.py
@@ -57,17 +57,3 @@ class ChipAndKickScreen(Screen):
 
         # Pass Wheel Screen parameters to super class
         super().__init__(lcd_display, screen_actions, actions)
-
-    def update_values(self, redis_dict):
-        """ Sync values with those from redis """
-        if not self.edit_mode:
-            for i in range(self.len):
-                if self.actions[i]["redis key"] == None:
-                    continue
-
-                if self.actions[i]["type"] == bool:
-                    self.actions[i]["value"] = (
-                        1 if redis_dict[self.actions[i]["redis key"]] else 0
-                    )
-                else:
-                    self.actions[i]["value"] = redis_dict[self.actions[i]["redis key"]]
